+++ SurviveCore/World/Generating/WorldBiomes/Ocean.cs
@@ -2,7 +2,7 @@
 
 namespace SurviveCore.World.Generating.WorldBiomes {
     public class OceanBiome : Biome {
-        public override void FillChunk(Chunk c, Random r, int x, int z, int height) {
+        public override void FillChunk(Chunk c, Random r, int x, int z, float height) {
             for (int y = 0; y < Chunk.Size; y++) {
                 if(c.Location.WY + y <= height - 1 - (height / 20))
                     c.SetBlockDirect(x, y, z, Blocks.Stone);
@@ -21,7 +21,7 @@ namespace SurviveCore.World.Generating.WorldBiomes {
         public OceanSelector(int p) : base(p) {
         }
 
-        public override Biome GetBiome(int x, int z, int height, float temperature, float humidity) {
+        public override Biome GetBiome(int x, int z, float height, float temperature, float humidity) {
             return height < AdvancedWorldGenerator.SeaLevel ? ocean : null;
         }
     }
